+++ docs/changes.rst
@@ -50,6 +50,8 @@ Pending
 * Stopped the history panel buttons from submitting their form when clicked
   before the panel script has loaded, which navigated away from the page.
 * Added support for Django 6.1.
+* Added open in editor functionality to templates panel using ``EDITOR``
+  setting.
 
 7.0.0 (2026-06-17)
 ------------------